requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0         # 用于读取.env文件
requests>=2.31.0  # 新增这一行